Delete messy/unused code, create sonar_image_processing.py file, and create functions for abstraction
