+++ onboard/catkin_ws/src/sonar/scripts/decodePingPythonPing360.py
@@ -16,6 +16,7 @@ from dataclasses import dataclass
 from typing import IO, Any, Set
 from localSonar import Sonar
 
+
 def indent(obj, by=' '*4):
     return by + str(obj).replace('\n', f'\n{by}')
 
@@ -270,105 +271,6 @@ class Ping360Settings:
         return v_sound * self.sample_period_us * 1e-6 / 2
 
 
-def increase_brightness(img, value=20):
-    hsv = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
-    h, s, v = cv2.split(hsv)
-
-    lim = 255 - value
-    v[v > lim] = 255
-    v[v <= lim] += value
-
-    final_hsv = cv2.merge((h, s, v))
-    img = cv2.cvtColor(final_hsv, cv2.COLOR_HSV2BGR)
-    return img
-
-def find_gate_posts(img): 
-    
-    greyscale_image = cv2.cvtColor(img.astype(np.uint8), cv2.COLOR_GRAY2BGR)
-    cm_image = cv2.applyColorMap(greyscale_image, cv2.COLORMAP_VIRIDIS)
-
-    cm_image = cv2.imread('onboard\\catkin_ws\\src\\sonar\\scripts\\sampleData\\Sonar_Image2.jpeg', cv2.IMREAD_COLOR)
-
-    cm_copy_image = cm_image
-    cv2.copyTo(cm_image, cm_copy_image)
-    cm_image = cv2.medianBlur(cm_image,5) # blur image
-
-    lower_color_bounds = (40,80,0) # filter out lower values (ie blue)
-    upper_color_bounds = (230,250,255) #filter out too high values
-    mask = cv2.inRange(cm_image,lower_color_bounds,upper_color_bounds)
-
-    cm_circles = cv2.findContours(mask, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)[-2]
-    cm_circles = sorted(cm_circles, key=cv2.contourArea, reverse=True)
-    cm_circles = list(filter(lambda x: (cv2.contourArea(x) > 200), cm_circles)) 
-    cm_circles = list(filter(lambda x: (cv2.arcLength(x, True)**2/(4*math.pi*cv2.contourArea(x)) < 5.4), cm_circles)) 
-
-    #return if not both goal posts found 
-    if(len(cm_circles) < 1):
-        print("not enough circles found")
-        return None
-
-    filtered_circles = cm_circles[0:2]
-
-    circle_positions = []
-    for circle in filtered_circles:  #find center of circle code
-        M = cv2.moments(circle)
-        cX = int(M["m10"] / M["m00"])
-        cY = int(M["m01"] / M["m00"])
-        circle_positions.append((cX,cY))
-        #print("object at " + "x: " + str(cX) + "  Y: " + str(cY)  +  " has circularity : " + str(perimeter**2/ (4*math.pi*area) ) )
-        #cv2.circle(cm_copy_image, (cX, cY), 3, (255, 255, 255), -1)
-
-    cv2.drawContours(cm_copy_image, filtered_circles, -1, (0,255,0), 2)
-    cv2.imshow("image", cm_copy_image)
-    cv2.waitKey(0)
-
-    return circle_positions
-
-
-def find_bouy(img): 
-    
-    greyscale_image = cv2.cvtColor(img.astype(np.uint8), cv2.COLOR_GRAY2BGR)
-    cm_image = cv2.applyColorMap(greyscale_image, cv2.COLORMAP_VIRIDIS)
-
-    cm_image = cv2.imread('onboard\\catkin_ws\\src\\sonar\\scripts\\sampleData\\Sonar_Image3.jpeg', cv2.IMREAD_COLOR)
-
-    cm_copy_image = cm_image
-    cv2.copyTo(cm_image, cm_copy_image)
-    cm_image = cv2.medianBlur(cm_image,5) # blur image
-
-    lower_color_bounds = (40,80,0) # filter out lower values (ie blue)
-    upper_color_bounds = (230,250,255) #filter out too high values
-    mask = cv2.inRange(cm_image,lower_color_bounds,upper_color_bounds)
-
-    cv2.imshow("image", mask)
-    cv2.waitKey(0)
-
-
-
-
-    cm_circles = cv2.findContours(mask, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)[-2]
-    cm_circles = list(filter(lambda x: (cv2.contourArea(x) > 100), cm_circles)) 
-
-    cm_circles = sorted(cm_circles, key=lambda x: (cv2.arcLength(x, True)**2/(4*math.pi*cv2.contourArea(x))), reverse=True)
-    #cm_circles = list(filter(lambda x: (cv2.arcLength(x, True)**2/(4*math.pi*cv2.contourArea(x)) < 5.4), cm_circles)) 
-
-
-
-    filtered_circles = cm_circles[0:1]
-
-    circle_positions = []
-    for circle in filtered_circles:  #find center of circle code
-        M = cv2.moments(circle)
-        cX = int(M["m10"] / M["m00"])
-        cY = int(M["m01"] / M["m00"])
-        circle_positions.append((cX,cY))
-        #print("object at " + "x: " + str(cX) + "  Y: " + str(cY)  +  " has circularity : " + str(perimeter**2/ (4*math.pi*area) ) )
-        #cv2.circle(cm_copy_image, (cX, cY), 3, (255, 255, 255), -1)
-
-    cv2.drawContours(cm_copy_image, filtered_circles, -1, (0,255,0), 2)
-    cv2.imshow("image", cm_copy_image)
-    cv2.waitKey(0)
-
 def getdecodedfile(localfilename):
     import os
     dirname = os.path.dirname(__file__)
@@ -379,6 +281,7 @@ def getdecodedfile(localfilename):
     log = PingViewerLogReader("\\sampleData\\"+filename)
 
     return log.parser()
+
 
 if __name__ == "__main__":
